requirements.txt
cffi==1.15.1
cryptography==38.0.1
DateTime==4.7
numpy==1.23.3
pycparser==2.21
pycryptodome==3.15.0
pyodbc==4.0.34
pytz==2022.2.1
zope.interface==5.4.0
mysql-connector-python==8.0.27
